Fix: Estandarizar ConcurrencyToken como string y revertir hacks en tests
- Se actualizó [WalletService.yaml](cci:7://file:///Users/polancou/RiderProjects/WalletService/Wallet.RestAPI/RestAPISpec/WalletService.yaml:0:0-0:0) eliminando `format: byte` de `concurrencyToken` para tratarlo puramente como string.
- Se agregaron las propiedades `ConcurrencyToken` faltantes en los modelos `*Result` (DTOs) para alinearse con la especificación.
- Se restauró la configuración de `AutoMapper` para convertir correctamente los tokens de `byte[]` a string Base64.
- Se revirtieron los tests de integración (como `ConcurrencyIntegrationTest` y `EmpresaApiTest`) para utilizar el token devuelto por la API, eliminando la dependencia directa de la base de datos para obtener este valor.
- Se validó la especificación con linter y se verificó que todos los tests pasan correctamente.

+++ Wallet.RestAPI/Models/ProductoUpdateRequest.cs
@@ -1,7 +1,7 @@
 /*
  * Wallet Service API
  *
- * Api para exponer la funcionalidad de wallet service.
+ * Api para exponer la funcionalidad de wallet service. 
  *
  * OpenAPI spec version: 0.1.0
  * Contact: edilberto_diaz14@hotmail.com
@@ -17,54 +17,59 @@ using Newtonsoft.Json;
 namespace Wallet.RestAPI.Models
 {
     /// <summary>
-    /// Estructura para crear/actualizar producto de proveedor
+    /// 
     /// </summary>
     [DataContract]
     public partial class ProductoUpdateRequest : IEquatable<ProductoUpdateRequest>
-    {
+    { 
         /// <summary>
         /// Gets or Sets Sku
         /// </summary>
         [Required]
+
         [MaxLength(50)]
-        [DataMember(Name = "sku")]
+        [DataMember(Name="sku")]
         public string Sku { get; set; }
 
         /// <summary>
         /// Gets or Sets Nombre
         /// </summary>
         [Required]
+
         [MaxLength(100)]
-        [DataMember(Name = "nombre")]
+        [DataMember(Name="nombre")]
         public string Nombre { get; set; }
 
         /// <summary>
         /// Gets or Sets Precio
         /// </summary>
 
-        [DataMember(Name = "precio")]
-        public decimal? Precio { get; set; }
+        [DataMember(Name="precio")]
+        public double? Precio { get; set; }
 
         /// <summary>
         /// Gets or Sets UrlIcono
         /// </summary>
         [Required]
+
         [MaxLength(255)]
-        [DataMember(Name = "urlIcono")]
+        [DataMember(Name="urlIcono")]
         public string UrlIcono { get; set; }
 
         /// <summary>
         /// Gets or Sets Categoria
         /// </summary>
         [Required]
-        [DataMember(Name = "categoria")]
+
+        [DataMember(Name="categoria")]
         public CategoriaEnum Categoria { get; set; }
 
         /// <summary>
         /// Gets or Sets ConcurrencyToken
         /// </summary>
         [Required]
-        [DataMember(Name = "concurrencyToken")]
+
+        [DataMember(Name="concurrencyToken")]
         public string ConcurrencyToken { get; set; }
 
         /// <summary>
@@ -116,32 +121,32 @@ namespace Wallet.RestAPI.Models
             if (ReferenceEquals(null, other)) return false;
             if (ReferenceEquals(this, other)) return true;
 
-            return
+            return 
                 (
                     Sku == other.Sku ||
                     Sku != null &&
                     Sku.Equals(other.Sku)
-                ) &&
+                ) && 
                 (
                     Nombre == other.Nombre ||
                     Nombre != null &&
                     Nombre.Equals(other.Nombre)
-                ) &&
+                ) && 
                 (
                     Precio == other.Precio ||
                     Precio != null &&
                     Precio.Equals(other.Precio)
-                ) &&
+                ) && 
                 (
                     UrlIcono == other.UrlIcono ||
                     UrlIcono != null &&
                     UrlIcono.Equals(other.UrlIcono)
-                ) &&
+                ) && 
                 (
                     Categoria == other.Categoria ||
                     Categoria != null &&
                     Categoria.Equals(other.Categoria)
-                ) &&
+                ) && 
                 (
                     ConcurrencyToken == other.ConcurrencyToken ||
                     ConcurrencyToken != null &&
@@ -159,25 +164,24 @@ namespace Wallet.RestAPI.Models
             {
                 var hashCode = 41;
                 // Suitable nullity checks etc, of course :)
-                if (Sku != null)
+                    if (Sku != null)
                     hashCode = hashCode * 59 + Sku.GetHashCode();
-                if (Nombre != null)
+                    if (Nombre != null)
                     hashCode = hashCode * 59 + Nombre.GetHashCode();
-                if (Precio != null)
+                    if (Precio != null)
                     hashCode = hashCode * 59 + Precio.GetHashCode();
-                if (UrlIcono != null)
+                    if (UrlIcono != null)
                     hashCode = hashCode * 59 + UrlIcono.GetHashCode();
-                if (Categoria != null)
+                    if (Categoria != null)
                     hashCode = hashCode * 59 + Categoria.GetHashCode();
-                if (ConcurrencyToken != null)
+                    if (ConcurrencyToken != null)
                     hashCode = hashCode * 59 + ConcurrencyToken.GetHashCode();
                 return hashCode;
             }
         }
 
         #region Operators
-
-#pragma warning disable 1591
+        #pragma warning disable 1591
 
         public static bool operator ==(ProductoUpdateRequest left, ProductoUpdateRequest right)
         {
@@ -189,8 +193,7 @@ namespace Wallet.RestAPI.Models
             return !Equals(left, right);
         }
 
-#pragma warning restore 1591
-
+        #pragma warning restore 1591
         #endregion Operators
     }
 }
